+++ src/instructable/server/dal/SingleInstance.java
@@ -25,9 +25,11 @@ public class SingleInstance
     private Map<String, FieldHolder> fields;
 
     long lastAccess; //not kept in DB
+    boolean connectToDB;
 
-    private SingleInstance(String userId, String conceptName, String instanceName, boolean mutable)
+    private SingleInstance(String userId, String conceptName, String instanceName, boolean mutable, boolean connectToDB)
     {
+        this.connectToDB = connectToDB;
         this.userId = userId;
         this.conceptName = conceptName;
         this.instanceName = instanceName;
@@ -35,10 +37,10 @@ public class SingleInstance
         lastAccess = System.nanoTime();
     }
 
-    static SingleInstance loadInstanceFieldsFromDB(String userId, String conceptName, String instanceName, boolean mutable, List<FieldDescription> fieldsInType)
+    static SingleInstance loadInstanceFieldsFromDB(String userId, String conceptName, String instanceName, boolean mutable, List<FieldDescription> fieldsInType, boolean connectToDB)
     {
         //TODO: must be called on initialization!!!
-        SingleInstance singleInstance = new SingleInstance(userId, conceptName, instanceName, mutable);
+        SingleInstance singleInstance = new SingleInstance(userId, conceptName, instanceName, mutable, connectToDB);
         singleInstance.loadFieldsFromDB(fieldsInType);
         return singleInstance;
     }
@@ -46,26 +48,29 @@ public class SingleInstance
     /**
      * instance shouldn't already exist in DB.
      */
-    static SingleInstance createNewInstance(String userId, String conceptName, String instanceName, boolean mutable, List<FieldDescription> fieldsInType)
+    static SingleInstance createNewInstance(String userId, String conceptName, String instanceName, boolean mutable, List<FieldDescription> fieldsInType, boolean connectToDB)
     {
-        //update DB!!!
-        try (
-                Connection connection = InstDataSource.getDataSource().getConnection();
-                PreparedStatement pstmt = connection.prepareStatement("insert into " + DBUtils.instancesTableName + " (" + DBUtils.userIdColName + "," + DBUtils.conceptColName + "," + DBUtils.instanceColName + "," + DBUtils.mutableColName + ") values (?,?,?,?)");
-        )
+        if (connectToDB)
         {
-            pstmt.setString(1, userId);
-            pstmt.setString(2, conceptName);
-            pstmt.setString(3, instanceName);
-            pstmt.setBoolean(4, mutable);
+            //update DB!!!
+            try (
+                    Connection connection = InstDataSource.getDataSource().getConnection();
+                    PreparedStatement pstmt = connection.prepareStatement("insert into " + DBUtils.instancesTableName + " (" + DBUtils.userIdColName + "," + DBUtils.conceptColName + "," + DBUtils.instanceColName + "," + DBUtils.mutableColName + ") values (?,?,?,?)");
+            )
+            {
+                pstmt.setString(1, userId);
+                pstmt.setString(2, conceptName);
+                pstmt.setString(3, instanceName);
+                pstmt.setBoolean(4, mutable);
 
-            pstmt.executeUpdate();
-        } catch (SQLException e)
-        {
-            e.printStackTrace();
+                pstmt.executeUpdate();
+            } catch (SQLException e)
+            {
+                e.printStackTrace();
+            }
         }
 
-        SingleInstance singleInstance = new SingleInstance(userId, conceptName, instanceName, mutable);
+        SingleInstance singleInstance = new SingleInstance(userId, conceptName, instanceName, mutable, connectToDB);
         singleInstance.createNewFields(fieldsInType);
         return singleInstance;
     }
@@ -85,54 +90,57 @@ public class SingleInstance
     private void loadFieldsFromDB(List<FieldDescription> fieldsInType)
     {
         fields = new HashMap<>();
-        //fill map from DB!         // reads all fields at once.
-        //TODO: check if works
-
-        try (
-                Connection connection = InstDataSource.getDataSource().getConnection();
-                PreparedStatement pstmt = connection.prepareStatement("select " + DBUtils.fieldColName + "," + DBUtils.fieldJSonValColName + " from " + DBUtils.instanceValTableName + " where " + DBUtils.userIdColName + "=?" + " and " + DBUtils.conceptColName + "=?" + " and " + DBUtils.instanceColName + "=?");
-        )
+        if (connectToDB)
         {
-            pstmt.setString(1, userId);
-            pstmt.setString(2, conceptName);
-            pstmt.setString(3, instanceName);
+            //fill map from DB!         // reads all fields at once.
+            //TODO: check if works
 
-            try (ResultSet resultSet = pstmt.executeQuery())
+            try (
+                    Connection connection = InstDataSource.getDataSource().getConnection();
+                    PreparedStatement pstmt = connection.prepareStatement("select " + DBUtils.fieldColName + "," + DBUtils.fieldJSonValColName + " from " + DBUtils.instanceValTableName + " where " + DBUtils.userIdColName + "=?" + " and " + DBUtils.conceptColName + "=?" + " and " + DBUtils.instanceColName + "=?");
+            )
             {
-                Map<String, Optional<JSONObject>> fieldsVals = new HashMap<>();
-                while (resultSet.next())
+                pstmt.setString(1, userId);
+                pstmt.setString(2, conceptName);
+                pstmt.setString(3, instanceName);
+
+                try (ResultSet resultSet = pstmt.executeQuery())
                 {
-                    String fieldResName = resultSet.getString(DBUtils.fieldColName);
-                    String jsonValAsStr = resultSet.getString(DBUtils.fieldJSonValColName);
-                    Optional<JSONObject> jsonValue = Optional.empty();
-                    try
+                    Map<String, Optional<JSONObject>> fieldsVals = new HashMap<>();
+                    while (resultSet.next())
                     {
-                        if (jsonValAsStr != null)
+                        String fieldResName = resultSet.getString(DBUtils.fieldColName);
+                        String jsonValAsStr = resultSet.getString(DBUtils.fieldJSonValColName);
+                        Optional<JSONObject> jsonValue = Optional.empty();
+                        try
                         {
-                            jsonValue = Optional.of(new JSONObject(jsonValAsStr));
+                            if (jsonValAsStr != null)
+                            {
+                                jsonValue = Optional.of(new JSONObject(jsonValAsStr));
+                            }
+                        } catch (JSONException e)
+                        {
+                            e.printStackTrace();
                         }
-                    } catch (JSONException e)
-                    {
-                        e.printStackTrace();
+                        fieldsVals.put(fieldResName, jsonValue);
                     }
-                    fieldsVals.put(fieldResName, jsonValue);
-                }
-                for (FieldDescription fieldDescription : fieldsInType)
-                {
-                    //need to create a FieldHolder also if has no JSon value
-                    String fieldName = fieldDescription.fieldName;
-                    Optional<JSONObject> jsonValue = Optional.empty();
-                    if (fieldsVals.containsKey(fieldName))
+                    for (FieldDescription fieldDescription : fieldsInType)
                     {
-                        jsonValue = fieldsVals.get(fieldName);
+                        //need to create a FieldHolder also if has no JSon value
+                        String fieldName = fieldDescription.fieldName;
+                        Optional<JSONObject> jsonValue = Optional.empty();
+                        if (fieldsVals.containsKey(fieldName))
+                        {
+                            jsonValue = fieldsVals.get(fieldName);
+                        }
+                        FieldHolder fieldHolder = new FieldHolder(fieldDescription, instanceName, new FieldChanged(this, fieldName), jsonValue);
+                        fields.put(fieldDescription.fieldName, fieldHolder);
                     }
-                    FieldHolder fieldHolder = new FieldHolder(fieldDescription, instanceName, new FieldChanged(this, fieldName), jsonValue);
-                    fields.put(fieldDescription.fieldName, fieldHolder);
                 }
+            } catch (SQLException e)
+            {
+                e.printStackTrace();
             }
-        } catch (SQLException e)
-        {
-            e.printStackTrace();
         }
     }
 
@@ -145,38 +153,41 @@ public class SingleInstance
     {
         String oldName = instanceName;
 
-        //update DB instance
-        try (
-                Connection connection = InstDataSource.getDataSource().getConnection();
-                PreparedStatement pstmt = connection.prepareStatement("update " + DBUtils.instancesTableName + " set " + DBUtils.instanceColName + " = ? where " + DBUtils.userIdColName + "=?" + " and " + DBUtils.conceptColName + "=?" + " and " + DBUtils.instanceColName + "=?");
-        )
+        if (connectToDB)
         {
-            pstmt.setString(1, newName);
-            pstmt.setString(2, userId);
-            pstmt.setString(3, conceptName);
-            pstmt.setString(4, oldName);
+            //update DB instance
+            try (
+                    Connection connection = InstDataSource.getDataSource().getConnection();
+                    PreparedStatement pstmt = connection.prepareStatement("update " + DBUtils.instancesTableName + " set " + DBUtils.instanceColName + " = ? where " + DBUtils.userIdColName + "=?" + " and " + DBUtils.conceptColName + "=?" + " and " + DBUtils.instanceColName + "=?");
+            )
+            {
+                pstmt.setString(1, newName);
+                pstmt.setString(2, userId);
+                pstmt.setString(3, conceptName);
+                pstmt.setString(4, oldName);
 
-            pstmt.executeUpdate();
-        } catch (SQLException e)
-        {
-            e.printStackTrace();
-        }
+                pstmt.executeUpdate();
+            } catch (SQLException e)
+            {
+                e.printStackTrace();
+            }
 
-        //update all fields in DB
-        try (
-                Connection connection = InstDataSource.getDataSource().getConnection();
-                PreparedStatement pstmt = connection.prepareStatement("update " + DBUtils.instanceValTableName + " set " + DBUtils.instanceColName + " = ? where " + DBUtils.userIdColName + "=?" + " and " + DBUtils.conceptColName + "=?" + " and " + DBUtils.instanceColName + "=?");
-        )
-        {
-            pstmt.setString(1, newName);
-            pstmt.setString(2, userId);
-            pstmt.setString(3, conceptName);
-            pstmt.setString(4, oldName);
+            //update all fields in DB
+            try (
+                    Connection connection = InstDataSource.getDataSource().getConnection();
+                    PreparedStatement pstmt = connection.prepareStatement("update " + DBUtils.instanceValTableName + " set " + DBUtils.instanceColName + " = ? where " + DBUtils.userIdColName + "=?" + " and " + DBUtils.conceptColName + "=?" + " and " + DBUtils.instanceColName + "=?");
+            )
+            {
+                pstmt.setString(1, newName);
+                pstmt.setString(2, userId);
+                pstmt.setString(3, conceptName);
+                pstmt.setString(4, oldName);
 
-            pstmt.executeUpdate();
-        } catch (SQLException e)
-        {
-            e.printStackTrace();
+                pstmt.executeUpdate();
+            } catch (SQLException e)
+            {
+                e.printStackTrace();
+            }
         }
 
 
@@ -208,21 +219,25 @@ public class SingleInstance
     public void fieldWasRemovedFromConcept(String fieldName)
     {
         fields.remove(fieldName);
-        //update the DB!!!
-        try (
-                Connection connection = InstDataSource.getDataSource().getConnection();
-                PreparedStatement pstmt = connection.prepareStatement("delete from " + DBUtils.instancesTableName + " where " + DBUtils.userIdColName + "=?" + " and " + DBUtils.conceptColName + "=?" + " and " + DBUtils.instanceColName + "=?" + " and " + DBUtils.fieldColName + "=?");
-        )
-        {
-            pstmt.setString(1, userId);
-            pstmt.setString(2, conceptName);
-            pstmt.setString(3, instanceName);
-            pstmt.setString(4, fieldName);
 
-            pstmt.executeUpdate();
-        } catch (SQLException e)
+        if (connectToDB)
         {
-            e.printStackTrace();
+            //update the DB!!!
+            try (
+                    Connection connection = InstDataSource.getDataSource().getConnection();
+                    PreparedStatement pstmt = connection.prepareStatement("delete from " + DBUtils.instancesTableName + " where " + DBUtils.userIdColName + "=?" + " and " + DBUtils.conceptColName + "=?" + " and " + DBUtils.instanceColName + "=?" + " and " + DBUtils.fieldColName + "=?");
+            )
+            {
+                pstmt.setString(1, userId);
+                pstmt.setString(2, conceptName);
+                pstmt.setString(3, instanceName);
+                pstmt.setString(4, fieldName);
+
+                pstmt.executeUpdate();
+            } catch (SQLException e)
+            {
+                e.printStackTrace();
+            }
         }
     }
 
@@ -244,57 +259,64 @@ public class SingleInstance
     public void changeMutability(boolean newMutability)
     {
         mutable = newMutability;
-        //update DB!!!
-        try (
-                Connection connection = InstDataSource.getDataSource().getConnection();
-                PreparedStatement pstmt = connection.prepareStatement("update " + DBUtils.instancesTableName + " set " + DBUtils.mutableColName + " = ? where " + DBUtils.userIdColName + "=?" + " and " + DBUtils.conceptColName + "=?" + " and " + DBUtils.instanceColName + "=?");
 
-        )
+        if (connectToDB)
         {
-            pstmt.setBoolean(1, newMutability);
-            pstmt.setString(2, userId);
-            pstmt.setString(3, conceptName);
-            pstmt.setString(4, instanceName);
+            //update DB!!!
+            try (
+                    Connection connection = InstDataSource.getDataSource().getConnection();
+                    PreparedStatement pstmt = connection.prepareStatement("update " + DBUtils.instancesTableName + " set " + DBUtils.mutableColName + " = ? where " + DBUtils.userIdColName + "=?" + " and " + DBUtils.conceptColName + "=?" + " and " + DBUtils.instanceColName + "=?");
 
-            pstmt.executeUpdate();
-        } catch (SQLException e)
-        {
-            e.printStackTrace();
+            )
+            {
+                pstmt.setBoolean(1, newMutability);
+                pstmt.setString(2, userId);
+                pstmt.setString(3, conceptName);
+                pstmt.setString(4, instanceName);
+
+                pstmt.executeUpdate();
+            } catch (SQLException e)
+            {
+                e.printStackTrace();
+            }
         }
     }
 
     void removeFromDB()
     {
-        //delete instance!!!
-        try (
-                Connection connection = InstDataSource.getDataSource().getConnection();
-                PreparedStatement pstmt = connection.prepareStatement("delete from " + DBUtils.instancesTableName + " where " + DBUtils.userIdColName + "=?" + " and " + DBUtils.conceptColName + "=?" + " and " + DBUtils.instanceColName + "=?");
-        )
+        if (connectToDB)
         {
-            pstmt.setString(1, userId);
-            pstmt.setString(2, conceptName);
-            pstmt.setString(3, instanceName);
+            //delete instance!!!
+            try (
+                    Connection connection = InstDataSource.getDataSource().getConnection();
+                    PreparedStatement pstmt = connection.prepareStatement("delete from " + DBUtils.instancesTableName + " where " + DBUtils.userIdColName + "=?" + " and " + DBUtils.conceptColName + "=?" + " and " + DBUtils.instanceColName + "=?");
+            )
+            {
+                pstmt.setString(1, userId);
+                pstmt.setString(2, conceptName);
+                pstmt.setString(3, instanceName);
 
-            pstmt.executeUpdate();
-        } catch (SQLException e)
-        {
-            e.printStackTrace();
-        }
+                pstmt.executeUpdate();
+            } catch (SQLException e)
+            {
+                e.printStackTrace();
+            }
 
-        //delete all fields that exist
-        try (
-                Connection connection = InstDataSource.getDataSource().getConnection();
-                PreparedStatement pstmt = connection.prepareStatement("delete from " + DBUtils.instanceValTableName + " where " + DBUtils.userIdColName + "=?" + " and " + DBUtils.conceptColName + "=?" + " and " + DBUtils.instanceColName + "=?");
-        )
-        {
-            pstmt.setString(1, userId);
-            pstmt.setString(2, conceptName);
-            pstmt.setString(3, instanceName);
+            //delete all fields that exist
+            try (
+                    Connection connection = InstDataSource.getDataSource().getConnection();
+                    PreparedStatement pstmt = connection.prepareStatement("delete from " + DBUtils.instanceValTableName + " where " + DBUtils.userIdColName + "=?" + " and " + DBUtils.conceptColName + "=?" + " and " + DBUtils.instanceColName + "=?");
+            )
+            {
+                pstmt.setString(1, userId);
+                pstmt.setString(2, conceptName);
+                pstmt.setString(3, instanceName);
 
-            pstmt.executeUpdate();
-        } catch (SQLException e)
-        {
-            e.printStackTrace();
+                pstmt.executeUpdate();
+            } catch (SQLException e)
+            {
+                e.printStackTrace();
+            }
         }
     }
